1.5.0
- Moves everything around
- Adds TypeScript
- Cuts down on unneeded files being published
- Adds build script to package.json
