--- index.d.ts
@@ -1,217 +1,0 @@
-export class Pagination {
-    paginationInfo: {
-        portal: any;
-        pageList: any;
-        buttonList: any;
-        pagination: any;
-    };
-    options: {
-        timeout: number;
-        replyMessage: boolean;
-        autoDelete: boolean;
-        privateReply: boolean;
-        authorIndependent: boolean;
-        ephemeral: boolean;
-        disabledButtons: boolean;
-        pageBuilderData: any;
-        buttonBuilderData: any;
-        autoButton: {
-            toggle: boolean;
-            deleteButton: boolean;
-        };
-        progressBar: {
-            toggle: boolean;
-            slider: string;
-            bar: string;
-        };
-        selectMenu: {
-            toggle: boolean;
-            useTitle: boolean;
-            labels: any;
-        };
-    };
-    /**
-     * Sets the used portal for the pagination
-     * @param {Message | Interaction} portal
-     * @param {{
-     *    interaction_ephemeral?: Boolean
-     * }} settings
-     * @returns {exports.Pagination}
-     */
-    setPortal(portal: Message | Interaction, settings?: {
-        interaction_ephemeral?: boolean;
-    }): Pagination;
-    /**
-     * Set the buttonList for the paginationY
-     * @param {ButtonBuilder[]} buttonList
-     * @returns {exports.Pagination}
-     */
-    setButtonList(buttonList: ButtonBuilder[]): Pagination;
-    /**
-     * Set the pageList for the pagination
-     * @param {EmbedBuilder[]} pageList
-     * @returns {exports.Pagination}
-     */
-    setPageList(pageList: EmbedBuilder[]): Pagination;
-    /**
-     * Run the pagination
-     * @returns {exports.Pagination}
-     */
-    paginate(): Pagination;
-    /**
-     * How many milliseconds your pagination will run for
-     * @param {Number} timeout
-     * @returns {exports.Pagination}
-     */
-    setTimeout(timeout: number): Pagination;
-    /**
-     * Allows you to enable and edit a progressBar for your pagination
-     * @param {{
-     *    slider?: String,
-     *    bar?: String
-     * }} settings
-     * @returns {exports.Pagination}
-     */
-    setProgressBar(settings?: {
-        slider?: string;
-        bar?: string;
-    }): Pagination;
-    /**
-     * Enables replyMessage for your pagination
-     * @returns {exports.Pagination}
-     */
-    enableReplyMessage(): Pagination;
-    /**
-     * Enables autoDelete for your pagination
-     * @returns {exports.Pagination}
-     */
-    enableAutoDelete(): Pagination;
-    /**
-     * Enables privateReply for your pagination
-     * @returns {exports.Pagination}
-     */
-    enablePrivateReply(): Pagination;
-    /**
-     * Enables authorIndependent for your pagination
-     * @returns {exports.Pagination}
-     */
-    enableAuthorIndependent(): Pagination;
-    /**
-     * Enables autoButton for your pagination
-     * @param {Boolean} deleteButton
-     * @returns {exports.Pagination}
-     */
-    enableAutoButton(deleteButton?: boolean): Pagination;
-    /**
-     * Enables selectMenu for your pagination
-     * @param {{
-     *    customLabels?: Array,
-     *    useTitles?: Boolean
-     * }} settings
-     * @returns {exports.Pagination}
-     */
-    enableSelectMenu(settings?: {
-        customLabels?: any[];
-        useTitles?: boolean;
-    }): Pagination;
-    /**
-     * Disables the buttons being disabled and re applied to the pagination after the timeout ends
-     * @returns {exports.Pagination}
-     */
-    disableDisabledButtons(): Pagination;
-    /**
-     * Allows you to use the pagination to create the pages for the pagination
-     * @param {[{
-     *    title?: String,
-     *    url?: String,
-     *    author?: {
-     *       name: String,
-     *       icon_url?: String,
-     *       url?: String
-     *    },
-     *    description?: String,
-     *    thumbnailUrl?: String,
-     *    fields?: [{
-     *       name: String,
-     *       value: String,
-     *       inline?: Boolean
-     *    }],
-     *    imageUrl?: String,
-     *    color?: 'Default'
-     *       |  'Random'
-     *       |  'White'
-     *       |  'Aqua'
-     *       |  'Green'
-     *       |  'Blue'
-     *       |  'Yellow'
-     *       |  'Purple'
-     *       |  'LuminousVividPink'
-     *       |  'Fuchsia'
-     *       |  'Gold'
-     *       |  'Orange'
-     *       |  'Red'
-     *       |  'Grey'
-     *       |  'Navy'
-     *       |  'DarkAqua'
-     *       |  'DarkGreen'
-     *       |  'DarkBlue'
-     *       |  'DarkPurple'
-     *       |  'DarkVividPink'
-     *       |  'DarkGold'
-     *       |  'DarkOrange'
-     *       |  'DarkRed'
-     *       |  'DarkGrey'
-     *       |  'DarkerGrey'
-     *       |  'LightGrey'
-     *       |  'DarkNavy'
-     *       |  'Blurple'
-     *       |  'Greyple'
-     *       |  'DarkButNotBlack'
-     *       |  'NotQuiteBlack'
-     * }]} buildData
-     * @returns {exports.Pagination}
-     */
-    createPages(buildData?: [
-        {
-            title?: string;
-            url?: string;
-            author?: {
-                name: string;
-                icon_url?: string;
-                url?: string;
-            };
-            description?: string;
-            thumbnailUrl?: string;
-            fields?: [
-                {
-                    name: string;
-                    value: string;
-                    inline?: boolean;
-                }
-            ];
-            imageUrl?: string;
-            color?: 'Default' | 'Random' | 'White' | 'Aqua' | 'Green' | 'Blue' | 'Yellow' | 'Purple' | 'LuminousVividPink' | 'Fuchsia' | 'Gold' | 'Orange' | 'Red' | 'Grey' | 'Navy' | 'DarkAqua' | 'DarkGreen' | 'DarkBlue' | 'DarkPurple' | 'DarkVividPink' | 'DarkGold' | 'DarkOrange' | 'DarkRed' | 'DarkGrey' | 'DarkerGrey' | 'LightGrey' | 'DarkNavy' | 'Blurple' | 'Greyple' | 'DarkButNotBlack' | 'NotQuiteBlack';
-        }
-    ]): Pagination;
-    /**
-     * Allows you to use the pagination to create the buttons for the pagination
-     * @param {[{
-     *    customId: String,
-     *    label?: String,
-     *    style: "Primary"
-     *       | "Secondary"
-     *       | "Success"
-     *       | "Danger",
-     *    emoji?: String
-     * }]} buildData
-     * @returns {exports.Pagination}
-     */
-    createButtons(buildData?: [
-        {
-            customId: string;
-            label?: string;
-            style: "Primary" | "Secondary" | "Success" | "Danger";
-            emoji?: string;
-        }
-    ]): Pagination;
-}
