+++ src/lib/MessagePagination.js
@@ -1,15 +1,15 @@
+// @ts-nocheck
 // Dependencies
-const { Interaction, ActionRowBuilder,  EmbedBuilder, ButtonBuilder } = require("discord.js");
+const { Message, ActionRowBuilder, EmbedBuilder, ButtonBuilder } = require("discord.js");
 const { DisabledButtonCreator } = require("../util/ButtonCreator");
 const { SelectMenuCreator, DisabledSelectMenuCreator } = require("../util/SelectMenuCreator");
 const { ProgressBarCreator } = require("../util/ProgressBarCreator");
 const { FilterCreator } = require("../util/FilterCreator");
 // Params
 /**
- * The interaction pagination
- *
+ * The Message pagination
  * @param {{
- *    portal: Interaction,
+ *    portal: Message,
  *    pageList: EmbedBuilder[],
  *    buttonList: ButtonBuilder[],
  *    pagination: null
@@ -20,8 +20,8 @@ const { FilterCreator } = require("../util/FilterCreator");
  *    autoDelete: Boolean,
  *    privateReply: Boolean,
  *    authorIndependent: Boolean,
- *    pageBuilderData: Array,
- *    buttonBuilderData: Array,
+ *    pageBuilderData: Array<Object>,
+ *    buttonBuilderData: Array<Object>,
  *    ephemeral: Boolean,
  *    disabledButtons: Boolean,
  *    autoButton: {
@@ -35,14 +35,14 @@ const { FilterCreator } = require("../util/FilterCreator");
  *    },
  *    selectMenu: {
  *       toggle: Boolean,
- *       labels: Array,
+ *       labels: Array<String>,
  *       useTitle: Boolean
  *    }
  * }} options
  * @returns {EmbedBuilder[]} The pagination
  */
-// Interaction pagination
-exports.InteractionPagination = async(paginationInfo, options) => {
+// Message pagination
+exports.MessagePagination = async(paginationInfo, options) => {
    try {
       // Set page number
       let pageNumber = 0;
@@ -53,16 +53,14 @@ exports.InteractionPagination = async(paginationInfo, options) => {
       };
       let pagination;
       if (options.privateReply) {
-         paginationInfo.portal.deferred ? await paginationInfo.portal.editReply("The reply has been sent privately") : await paginationInfo.portal.reply("The reply has been sent privately");
-         pagination = await paginationInfo.portal.client.users.cache.get(paginationInfo.portal.member.user.id).send(paginationContent);
-      } else if (paginationInfo.portal.deferred) {
-         pagination = await paginationInfo.portal.editReply(paginationContent);
+         await paginationInfo.portal.channel.send("The reply has been sent privately");
+         pagination = await paginationInfo.portal.author.send(paginationContent);
       } else {
-         pagination = await paginationInfo.portal.reply(paginationContent);
+         pagination = options.replyMessage ? await paginationInfo.portal.reply(paginationContent) : await paginationInfo.portal.channel.send(paginationContent);
       }
       // Create collector
-      const collector = await pagination.createMessageComponentCollector({
-         filter: FilterCreator({interaction: paginationInfo.portal, buttonList: paginationInfo.buttonList, authorIndependent: options.authorIndependent, selectMenu: options.selectMenu}),
+      const collector = pagination.createMessageComponentCollector({
+         filter: FilterCreator({message: paginationInfo.portal, buttonList: paginationInfo.buttonList, authorIndependent: options.authorIndependent, selectMenu: options.selectMenu}),
          time: options.timeout
       });
       // Button inputs
@@ -104,14 +102,15 @@ exports.InteractionPagination = async(paginationInfo, options) => {
                   return;
                // Catch null
                case null:
-                  throw new Error("InteractionPagination ERROR: The button input is null");
+                  throw new Error("MessagePagination ERROR: The button input is null");
                // Catch undefined
                case undefined:
-                  throw new Error("InteractionPagination ERROR: The button input is undefined");
+                  throw new Error("MessagePagination ERROR: The button input is undefined");
                // Default
                default:
                   break;
             }
+            // Deferrer update
             if (!i.deferred) await i.deferUpdate();
             // Edit page after input
             await i.editReply({
@@ -124,7 +123,7 @@ exports.InteractionPagination = async(paginationInfo, options) => {
             // Deferrer update
             if (!i.deferred) await i.deferUpdate();
             // Edit page after input
-            pageNumber = i.values[0] - 1 ;
+            pageNumber = i.values[0] - 1;
             await i.editReply({
                embeds: [paginationInfo.pageList[pageNumber].setFooter({text: `${options.progressBar.toggle ? `${await ProgressBarCreator(paginationInfo.pageList.length, pageNumber, options.progressBar)}` : `Page ${pageNumber + 1} / ${paginationInfo.pageList.length}`}`})],
                fetchReply: true
@@ -137,22 +136,28 @@ exports.InteractionPagination = async(paginationInfo, options) => {
       collector.once("end", async() => {
          try {
             // Make sure the embed exists
-            await paginationInfo.portal.channel.messages.fetch({message: pagination.id});
+            await paginationInfo.portal.channel.messages.fetch(pagination.id);
             // Delete if autoDelete in enabled
-            if (options.autoDelete) return pagination.delete();
-            // No disabled buttons
-            if (!options.disabledButtons) return paginationInfo.portal.editReply({ components: [await DisabledSelectMenuCreator(pagination.components[0])] });
-            // Disable select menu
-            if (options.selectMenu.toggle) {
-               try {
-                  return paginationInfo.portal.editReply({ components: [await DisabledSelectMenuCreator(pagination.components[0])] });
-               } catch(error) {}
+            if (options.autoDelete) {
+               return pagination.delete();
             }
-            // Disable buttons
+            // No disabled buttons
+            else if (!options.disabledButtons) {
+               return pagination.edit({ components: [] });
+            }
             else {
-               try {
-                  return paginationInfo.portal.editReply({ components: [await DisabledButtonCreator(paginationInfo.buttonList)] });
-               } catch (error) {return}
+               // Disable select menu
+               if (options.selectMenu.toggle) {
+                  try {
+                     return pagination.edit({ components: [await DisabledSelectMenuCreator(pagination.components[0])] });
+                  } catch(error) {}
+               }
+               // Disable buttons
+               else {
+                  try {
+                     return pagination.edit({ components: [await DisabledButtonCreator(paginationInfo.buttonList)] });
+                  } catch (error) {return}
+               }               
             }
          } catch(error) {return}
       });
