+++ src/lib/PaginationBase.js
@@ -1,3 +1,4 @@
+// @ts-nocheck
 // Dependencies
 const { MessagePayload, EmbedBuilder } = require("discord.js");
 const { InteractionPagination } = require("./InteractionPagination");
