+++ src/util/PageCreator.js
@@ -1,3 +1,4 @@
+// @ts-nocheck
 // Dependencies
 const { EmbedBuilder } = require("discord.js");
 /**
