--- util/FilterCreator.d.ts
@@ -1,11 +1,0 @@
-export function FilterCreator({ message, interaction, buttonList, authorIndependent, selectMenu }: {
-    message?: Message;
-    interaction?: Interaction;
-    buttonList: ButtonBuilder;
-    authorIndependent: boolean;
-    selectMenu: {
-        toggle: boolean;
-        labels: any[];
-        useTitle: boolean;
-    };
-}): Function;
