--- util/ProgressBarCreator.d.ts
@@ -1,5 +1,0 @@
-export function ProgressBarCreator(pageListLength: number, pageNumber: number, progressBar: {
-    toggle: boolean;
-    slider?: string;
-    bar?: string;
-}): Promise<string>;
